chat-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
mcp[cli]==1.20.0
anthropic==0.43.0
google-genai>=1.0.0
